Made create class a component itself
